extract data-source into a component
