+++ src/app/main/result/item.result.component.ts
@@ -163,24 +163,9 @@ export class ItemResultComponent implements OnInit {
         this.dataSource = [...this.rows];
     }
 
-    selectDataSource(isPersonal: boolean) {
-        if (this.penguinService.isPersonal === isPersonal) {
-            return;
-        }
-        if (!window.localStorage) {
-            this._snackBar.open("您的浏览器暂不支持本地数据，请升级或者换浏览器再试。", "x", { duration: 2000 });
-            return;
-        }
-        let localStageTimesStr = localStorage.getItem("stageTimes");
-        let localDropMatrixStr = localStorage.getItem("dropMatrix");
-        if (isPersonal && (!localStageTimesStr || !localDropMatrixStr)) {
-            this._snackBar.open("您当前还未上传过掉落数据。", "x", { duration: 2000 });
-            return;
-        }
-        this.penguinService.isPersonal = isPersonal;
+    onRefreshResult() {
         if (this.selectedService.selections.result_by_item.selectedItem) {
             this._refreshItemResult();
         }
     }
-
 }
