+++ src/app/main/result/stage.result.component.html
@@ -1,19 +1,7 @@
 <div class='body'>
   <div class="row">
     <div class="col-lg-12">
-      <div class="row d-flex justify-content-end align-items-center">
-        <div class="source-label">数据来源：</div>
-        <div id="result-source-buttons">
-          <button type="button" class="btn btn-outline-danger" [ngClass]="{'active': !penguinService.isPersonal}"
-            (click)="selectDataSource(false)">
-            全平台
-          </button>
-          <button type="button" class="btn btn-outline-danger" [ngClass]="{'active': penguinService.isPersonal}"
-            (click)="selectDataSource(true)">
-            本地
-          </button>
-        </div>
-      </div>
+      <app-data-source (refreshResult)=onRefreshResult()></app-data-source>
       <app-stage-selector [chapter]="selectedService.selections.result_by_stage.selectedChapter"
         [stage]="selectedService.selections.result_by_stage.selectedStage" (chapterChange)="onChapterChange($event)"
         (stageChange)="onStageChange($event)" [chapterFilter]="resultStageFilter">
