+++ src/app/main/result/stage.result.component.ts
@@ -168,24 +168,10 @@ export class StageResultComponent implements OnInit {
         this.dataSource = [...this.rows];
     }
 
-    selectDataSource(isPersonal: boolean) {
-        if (this.penguinService.isPersonal === isPersonal) {
-            return;
-        }
-        if (!window.localStorage) {
-            this._snackBar.open("您的浏览器暂不支持本地数据，请升级或者换浏览器再试。", "x", { duration: 2000 });
-            return;
-        }
-        let localStageTimesStr = localStorage.getItem("stageTimes");
-        let localDropMatrixStr = localStorage.getItem("dropMatrix");
-        if (isPersonal && (!localStageTimesStr || !localDropMatrixStr)) {
-            this._snackBar.open("您当前还未上传过掉落数据。", "x", { duration: 2000 });
-            return;
-        }
-        this.penguinService.isPersonal = isPersonal;
-        if (this.selectedService.selections.result_by_stage.selectedStage && this.selectedService.selections.result_by_stage.selectedChapter) {
+    onRefreshResult() {
+        if (this.selectedService.selections.result_by_stage.selectedStage
+            && this.selectedService.selections.result_by_stage.selectedChapter) {
             this._refreshStageResult();
         }
     }
-
 }
